[chore] search -> explore로 디렉토리 변경, search, recommend를 해당 패키지의 하위 디렉토리로 변경
