+++ src/main/java/com/wayble/server/explore/dto/recommend/WaybleZoneRecommendResponseDto.java
@@ -1,0 +1,8 @@
+package com.wayble.server.explore.dto.recommend;
+
+public record WaybleZoneRecommendResponseDto(
+
+        String username
+
+) {
+}
