+++ src/main/java/com/wayble/server/explore/dto/search/SearchSliceDto.java
@@ -1,0 +1,9 @@
+package com.wayble.server.explore.dto.search;
+
+import java.util.List;
+
+public record SearchSliceDto<T>(
+        List<T> content,
+        boolean hasNext
+) {
+}
