--- src/main/java/com/wayble/server/recommend/dto/WaybleZoneRecommendResponseDto.java
@@ -1,8 +1,0 @@
-package com.wayble.server.recommend.dto;
-
-public record WaybleZoneRecommendResponseDto(
-
-        String username
-
-) {
-}
