--- src/main/java/com/wayble/server/recommend/repository/WaybleZoneRecommendRepository.java
@@ -1,7 +1,0 @@
-package com.wayble.server.recommend.repository;
-
-import com.wayble.server.search.entity.WaybleZoneDocument;
-import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
-
-public interface WaybleZoneRecommendRepository extends ElasticsearchRepository<WaybleZoneDocument, Long>, WaybleZoneSearchRepositoryCustom {
-}
