--- src/main/java/com/wayble/server/recommend/repository/WaybleZoneSearchRepositoryCustom.java
@@ -1,9 +1,0 @@
-package com.wayble.server.recommend.repository;
-
-import com.wayble.server.recommend.dto.WaybleZoneRecommendResponseDto;
-import com.wayble.server.user.entity.User;
-
-public interface WaybleZoneSearchRepositoryCustom {
-
-    WaybleZoneRecommendResponseDto searchPersonalWaybleZone(User user);
-}
