--- src/main/java/com/wayble/server/search/dto/SearchSliceDto.java
@@ -1,9 +1,0 @@
-package com.wayble.server.search.dto;
-
-import java.util.List;
-
-public record SearchSliceDto<T>(
-        List<T> content,
-        boolean hasNext
-) {
-}
