--- src/main/java/com/wayble/server/search/dto/WaybleZoneDocumentRegisterDto.java
@@ -1,17 +1,0 @@
-package com.wayble.server.search.dto;
-
-import com.wayble.server.common.entity.Address;
-import com.wayble.server.wayblezone.entity.WaybleZoneType;
-import lombok.Builder;
-
-@Builder
-public record WaybleZoneDocumentRegisterDto(
-        Long zoneId,
-        String zoneName,
-        WaybleZoneType waybleZoneType,
-        String thumbnailImageUrl,
-        Address address,
-        Double averageRating,
-        Long reviewCount
-) {
-}
